Moved db initialization to another file
